refactor: Modularize Flask app into 'project' package
- Create a 'project' package to hold application modules.
- Move route handlers from `app.py` to `project/routes.py`.
- Move helper functions `generate_plots` and `generate_html_table` to `project/routes.py`.
- Consolidate financial calculation logic by removing duplicated functions
  (`simulate_final_balance`, `find_required_portfolio`, `find_max_annual_expense`)
  from `app.py` and ensuring `project/routes.py` uses the versions from `financial_calcs.py`.
- Move `financial_calcs.py` and `constants.py` into the `project` package.
- Update all import statements to use relative imports within the `project` package
  (e.g., `from .financial_calcs import ...`) and adjust imports from the root `app` object.
- Refactor `app.py` to primarily handle app initialization, configuration,
  logging, and importing routes from the `project` package.
- Update test files (`tests/test_app.py`, `tests/test_finance_core.py`) to correctly
  import modules and patch objects from their new locations within the `project` package.
